+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project are documented in this file.
 The format adheres to Semantic Versioning.
 
+## 0.1.1 (2017-10-21)
+
+Features
+- Created class stub files: porfolio.py and project.py
+- Created system.py file for module-wide variables and functions.
+- Coded class decorator `applyToAllMethods()` to apply a function across all methods of a class.
+- Coded function decorator `logUsage()` to timestamp usage and measure elapsed time.
+
+
 ## 0.1.0 (2017-10-17)
 
 Features
@@ -11,6 +20,6 @@ Features
   Heavily based on Optima HIV/TB.
 - Created _version.py file to track version numbers and dates.
   Must be manually updated.
-- Created __init__.py file for the module.
+- Created \_\_init\_\_.py file for the module.
   Heavily based on Optima HIV/TB.
 - Added files for documentation: README.md and CHANGELOG.md